Parse the = character specially
The = mnemonic is an alias for the SET mnemonic; it sets a variable.
Some people like to write variable assignments without spaces, like
this:

a=0

Teach the parser to handle the = character in such a way that this
example parses as expected, i.e. a label (a), a mnemonic (=) and an
operand (0).