Refactoring in preparation for incremental re-parsing
Besides the refactoring, this commit also changes some behavior. Now,
mnemonics that identify block directives (such as ELSE) can no longer
refer to macros when they are in an invalid context. That's because the
LogicalLine is now always wrapped in a BlockDirectiveLine when the
mnemonic is a known block directive. This will facilitate incremental
re-parsing (no need to wrap/unwrap BlockDirectiveLines).