Allow more symbols in identifiers
Allow the quotation mark, the number sign, the dollar sign and the
apostrophe in identifiers, except as the first code point in an
identifier. The logical line parser learned to properly handle quotation
marks and apostrophes in identifiers (i.e. it learned not to make them
string delimiters in this context).

This commit also includes a bit of refactoring: the Identifier class has
been renamed to Syntax and now contains various other methods that
classify code points.