Remove implicit nodes
They provide no value (they couldn't be assembled because they were
children of a node with a parse error) and they cause a subtle (but
mostly harmless) issue with re-parsing
(ReparserSourceNodeProducer.next() crashes with a ClassCastException on
the RECYCLE_AFTER step if it is called after reparsing is complete).
Since there is no sensible way in which the re-parser can handle them
(without specific knowledge about them), the best way to resolve this
issue is to remove them completely.