Support operand packs in macros
A macro can not include "..." in the parameter list to designate an
operand pack. The operand pack can be located anywhere in the parameter
list, but there can be at most one. On the macro definition, if the
operand pack is preceded by x operands and followed by y operands, then
the pack includes all operands except the first x ones and the last y
ones.